IGAPP-636: Refactor and fix useNavigateToDeepLink
